🚚 Move python configs
