+++ external/anomaly/configs/utils/configuration_enums.py
