+++ external/anomaly/configs/utils/padim/__init__.py
